Minor API changes and fixes: - fix transaction management on JVM - rename GeoCoordinates to GeoBounds - rename Field and field builder methods: int to long and real to double - getOrCreateEdge/Node does not require a separate schema argument - additional method "updateField" which takes a lambda to produce a new value based on an old value

+++ src/commonMain/kotlin/pl/makenika/graphlite/sql/SqlContentValues.kt
@@ -19,6 +19,6 @@ package pl.makenika.graphlite.sql
 expect class SqlContentValues() {
     fun put(key: String, value: ByteArray?)
     fun put(key: String, value: Double?)
-    fun put(key: String, value: Int?)
+    fun put(key: String, value: Long?)
     fun put(key: String, value: String?)
 }
